backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.59.3
chromadb==0.5.23
neo4j==5.27.0
pandas==2.2.3
pdfplumber==0.11.4
python-pptx==1.0.2
olefile==0.47
tiktoken==0.8.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
sse-starlette==2.2.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
